api/requirements.txt
flask
connexion
uwsgi
redis

celery

requests
requests-cache
fake-useragent

pyyaml

numpy
scipy
pandas
networkx
